+++ apps/app/src/test/scala/com/daml/network/integration/tests/WalletTxLogTimeBasedIntegrationTest.scala
@@ -210,6 +210,7 @@ class WalletTxLogTimeBasedIntegrationTest
               Duration.ofMinutes(5),
             ),
           ),
+          getLedgerTime,
         ),
       )(
         "Charlie has received the CC",